refactor namespaces and source organisation
#20
