+++ CmakeLists.txt
@@ -10,18 +10,26 @@ set(CMAKE_RUNTIME_OUTPUT_DIRECTORY ./bin)
 
 add_library(engine STATIC engine/engine_gl.cpp engine/engine.cpp engine/glad.c engine/stb_image.cpp)
 target_include_directories(engine PUBLIC ./engine)
+set_property(TARGET engine PROPERTY CXX_STANDARD 17)
+set_property(TARGET engine PROPERTY CXX_STANDARD_REQUIRED ON)
 if (CMAKE_CXX_COMPILER_ID STREQUAL "GNU")
     target_link_libraries(engine -lmingw32 -lSDL2main -lSDL2)
 endif()
 
-add_library(game STATIC source/content/menu.cpp source/content/teststage.cpp source/content/player.cpp source/content/bullet.cpp source/content/enemy.cpp source/content/script/test/testbullet.cpp)
-target_include_directories(game PUBLIC . ./engine ./source/content)
+add_library(game STATIC source/game/bullet.cpp source/game/enemy.cpp source/game/menu.cpp source/game/player.cpp)
+target_include_directories(game PUBLIC . ./engine ./source)
 set_property(TARGET game PROPERTY CXX_STANDARD 17)
 set_property(TARGET game PROPERTY CXX_STANDARD_REQUIRED ON)
 
-add_executable(thszn source/main.cpp source/game.cpp)
+add_library(script STATIC source/script/test/testbullet.cpp source/script/test/teststage.cpp)
+target_include_directories(script PUBLIC . ./engine ./source ./source/game ./source/script/test)
+set_property(TARGET script PROPERTY CXX_STANDARD 17)
+set_property(TARGET script PROPERTY CXX_STANDARD_REQUIRED ON)
+target_link_libraries(script game)
+
+add_executable(thszn source/main.cpp source/state.cpp)
 target_include_directories(thszn PUBLIC .)
-target_link_libraries(thszn engine game)
+target_link_libraries(thszn engine game script)
 if (CMAKE_CXX_COMPILER_ID STREQUAL "GNU")
     target_link_libraries(thszn -Wl,-subsystem,console)
 endif()
