+++ engine/engine_gl.hpp
@@ -9,7 +9,6 @@
 #include <string>
 #include <fstream>
 #include <sstream>
-#include <iostream>
 
 #include "glm/glm.hpp"
 #include "glm/gtc/matrix_transform.hpp"
