--- source/content/script/test/testbullet.hpp
@@ -1,11 +1,0 @@
-#ifndef _TESTBULLET_H
-#define _TESTBULLET_H
-
-#include "..\..\bullet.hpp"
-
-namespace game::script::test {
-    void testbulletfunc(game::content::bullet_s*);
-}
-
-
-#endif
