+++ source/script/test/testbullet.hpp
@@ -1,0 +1,11 @@
+#ifndef _TESTBULLET_H
+#define _TESTBULLET_H
+
+#include "bullet.hpp"
+
+namespace Game::Script::Test {
+    void testbulletfunc(Game::bullet_s*);
+}
+
+
+#endif
